Move Paginators constants to Paginator
